+++ src/app/lib/actions.js
@@ -3,7 +3,7 @@
 import { cookies } from "next/headers";
 import { revalidatePath } from "next/cache";
 
-export async function createArticle(tags, state, formData) {
+export async function createArticle(tags, slug, state, formData) {
   try {
     const response = await fetch(`http://api:3000/api/articles`, {
       method: "POST",
@@ -27,7 +27,7 @@ export async function createArticle(tags, state, formData) {
 }
 
 export async function updateArticle(tags, slug, state, formData) {
-  console.log(tags);
+  console.log(tags, slug, formData);
   try {
     const response = await fetch(`http://api:3000/api/articles/${slug}`, {
       method: "PUT",
